+++ storage/store/full_kv.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
+	"time"
 
 	"github.com/streamingfast/bstream"
 	"github.com/streamingfast/derr"
@@ -52,6 +53,7 @@ func (s *FullKV) QuickLoad(ctx context.Context, atBlock bstream.BlockRef) error 
 		return ErrNoQuickSaveStore
 	}
 
+	start := time.Now()
 	filename := atBlock.ID() + ".quicksave"
 	s.logger.Debug("loading full store state from temporary file", zap.String("fileName", filename), zap.String("module_hash", s.moduleHash), zap.Uint64("block_num", atBlock.Num()), zap.String("block_id", atBlock.ID()))
 
@@ -73,6 +75,7 @@ func (s *FullKV) QuickLoad(ctx context.Context, atBlock bstream.BlockRef) error 
 		zap.Uint64("data_size", s.totalSizeBytes),
 		zap.Uint64("block_num", atBlock.Num()),
 		zap.String("block_id", atBlock.ID()),
+		zap.Duration("load_duration", time.Since(start)),
 	)
 	return nil
 }
@@ -81,12 +84,17 @@ func (s *FullKV) QuickSave(ctx context.Context, atBlockHash string) error {
 	if s.quickSaveStore == nil {
 		return ErrNoQuickSaveStore
 	}
+	start := time.Now()
 	s.logger.Info("quicksave: writing temporary store state", zap.Object("store", s))
 
 	store := s.quickSaveStore
 	filename := atBlockHash + ".quicksave"
 
-	reader := s.marshaller.MarshalStreamIter(s.kvImpl.Iter, nil, int64(s.totalSizeBytes))
+	snap, err := s.kvImpl.Snapshot()
+	if err != nil {
+		return fmt.Errorf("snapshotting store %q: %w", s.name, err)
+	}
+	reader := s.marshaller.MarshalStreamSnapshot(snap, nil)
 
 	fw := &fileWriter{
 		store:    store,
@@ -94,7 +102,12 @@ func (s *FullKV) QuickSave(ctx context.Context, atBlockHash string) error {
 		reader:   reader,
 	}
 
-	return fw.Write(ctx)
+	if err := fw.Write(ctx); err != nil {
+		return err
+	}
+
+	s.logger.Info("quicksave: temporary store state written", zap.String("fileName", filename), zap.Int("key_count", s.kvImpl.KeyCount()), zap.Uint64("data_size", s.totalSizeBytes), zap.Duration("save_duration", time.Since(start)))
+	return nil
 }
 
 var ErrInvalidFullKVFile = errors.New("unmarshal store error") // this error will bubble up to the user
@@ -139,7 +152,11 @@ func (s *FullKV) Save(endBoundaryBlock uint64) (*FileInfo, *fileWriter, error) {
 
 	file := NewCompleteFileInfo(s.name, s.moduleInitialBlock, endBoundaryBlock)
 
-	reader := s.marshaller.MarshalStreamIter(s.kvImpl.Iter, nil, int64(s.totalSizeBytes))
+	snap, err := s.kvImpl.Snapshot()
+	if err != nil {
+		return nil, nil, fmt.Errorf("snapshotting store %q: %w", s.name, err)
+	}
+	reader := s.marshaller.MarshalStreamSnapshot(snap, nil)
 
 	fw := &fileWriter{
 		store:    s.objStore,